FF-2251: Updated Revisions Explorer to use new modal system
