+++ Client/Shared/MainLayout/MainLayout.razor
@@ -51,7 +51,6 @@
 @if (Blocker != null)
 {
     <CascadingValue Value="Blocker" IsFixed="true">
-        <RevisionExplorer/>
         <AuditHistory/>
         <AuditEntryViewer/>
     </CascadingValue>
